feat: implement sequelize product, supplier

+++ services/product.service.js
@@ -1,8 +1,9 @@
 import ProductRepository from "../repositories/product.repository.js";
 import SupplierRepository from "../repositories/supplier.repository.js";
+import SaleRepository from "../repositories/sale.repository.js";
 
 const createProduct = async (product) => {
-  if (await SupplierRepository.getSupplier(product.supplier_id)) {
+  if (await SupplierRepository.getSupplier(product.supplierId)) {
     return await ProductRepository.insertProduct(product);
   }
   throw new Error("Supplier not found");
@@ -17,11 +18,14 @@ const getProduct = async (id) => {
 };
 
 const deleteProduct = async (id) => {
+  if (await SaleRepository.getSalesByProductId(id)) {
+    throw new Error("Product has sales");
+  }
   return await ProductRepository.deleteProduct(id);
 };
 
 const updateProduct = async (product) => {
-  if (await SupplierRepository.getSupplier(product.supplier_id)) {
+  if (await SupplierRepository.getSupplier(product.supplierId)) {
     return await ProductRepository.updateProduct(product);
   }
   throw new Error("Supplier not found");
